// apps/api/generated/prisma/enums.ts
export const ExportFormat = {
  PPTX: 'PPTX',
  PDF: 'PDF',
  GOOGLE_SLIDES: 'GOOGLE_SLIDES',
  REVEAL_JS: 'REVEAL_JS',
  FIGMA: 'FIGMA'
} as const

export type ExportFormat = (typeof ExportFormat)[keyof typeof ExportFormat]


export const JobStatus = {
  QUEUED: 'QUEUED',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  PENDING_RETRY: 'PENDING_RETRY'
} as const

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus]